`Streamable` in terms of parsers.
