+++ src/Trisagion/Parsers/Streamable.hs
@@ -6,12 +6,9 @@ Parsers with @'Streamable' m a s@ constraints.
 
 module Trisagion.Parsers.Streamable (
     -- * Error types.
-    InputError (..),
     ValidationError (..),
 
     -- * Parsers @'Streamable' m a s => 'ParserT' m s e a@.
-    eoi,
-    one,
     skipOne,
     peek,
     satisfy,
@@ -25,22 +22,13 @@ import Data.Functor (($>))
 import Data.Functor.Identity (Identity (..))
 import Data.Void (Void)
 
--- Libraries.
-import Control.Monad.State (MonadState (..))
-import Control.Monad.Trans (MonadTrans (..))
-
 -- Package.
 import Trisagion.Types.Either ((:+:))
 import Trisagion.Types.ParseError (ParseError (..))
 import Trisagion.Typeclasses.HasOffset (HasOffset (..))
-import Trisagion.Typeclasses.Streamable (Streamable (..))
-import Trisagion.ParserT (ParserT, throw)
+import Trisagion.Typeclasses.Streamable (Streamable (..), InputError (..))
+import Trisagion.ParserT (ParserT, lookAhead, try)
 import Trisagion.Parsers.ParseError (validate)
-
-
-{- | The t'InputError' error type. -}
-newtype InputError = InputError Word
-    deriving stock (Eq, Show)
 
 
 {- | The t'ValidationError' error tag type thrown on failed validations. -}
@@ -49,38 +37,15 @@ newtype ValidationError e = ValidationError e
     deriving (Applicative, Monad) via Identity
 
 
-{- | Monadic check for nullity of the input stream. -}
-{-# INLINE eoi #-}
-eoi :: Streamable m a s => ParserT s Void m Bool
-eoi = get >>= \ xs -> lift (nullM xs)
-
-{- | Parse one element from the input stream. -}
-{-# INLINE one #-}
-one :: (Streamable m a s, HasOffset m s) => ParserT s (ParseError InputError) m a
-one = do
-    e <- get >>= \ xs -> lift ((flip ParseError (InputError 1)) <$> offset xs)
-    r <- get >>= \ xs -> lift (unconsM xs)
-    case r of
-        Nothing      -> throw e
-        Just (x, ys) -> put ys $> x
-
 {- | Skip one element from the input stream. -}
 {-# INLINE skipOne #-}
 skipOne :: Streamable m a s => ParserT s Void m ()
-skipOne = do
-    r <- get >>= \ xs -> lift (unconsM xs)
-    case r of
-        Nothing      -> pure ()
-        Just (_, ys) -> put ys
+skipOne = try one $> ()
 
 {- | Extract the first element from the input stream but without consuming input. -}
 {-# INLINE peek #-}
 peek :: Streamable m a s => ParserT s Void m (Maybe a)
-peek = do
-    r <- get >>= \ xs -> lift (unconsM xs)
-    case r of
-        Nothing     -> pure $ Nothing
-        Just (x, _) -> pure $ Just x
+peek = either (const Nothing) Just <$> lookAhead one
 
 {- | Parse one element from the input stream satisfying a predicate. -}
 {-# INLINE satisfy #-}
